requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
langchain==0.2.6
langchain-openai==0.1.14
openai==1.35.7
youtube-transcript-api==1.1.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis==5.0.6
upstash-redis==1.1.0
gumloop==1.0.0